Extract utils functions from Documents resolver to the utils folder
